This patch does four things: 1. Moves B/E event parsing from TraceEventImporter to a TimelineSliceGroup. 2. Removes the awkward concept of "secondary traces" and instead creates a generalized "importTraces" call 3. Stops the TimelineModel from "bloating" the model range and instead does that bloating in the viewport. This improves test clarity. 4. Makes the tests use addHTMLOutput for their output, so failures show up in a more reasonable place
The biggest change, #1, moves the processing of B/E events, including their autoclosing, subslice-creation and nesting-detection logic from TraceEventImporter into a base class, TimelineSliceGroup. This is needed because we are seeing more and more uses of B/E style events in the LinuxPerfImporter. This avoids duplication of that code.

One key change in the design as part of this move is that subRows is
now built on-demand, based purely on the timestamps of the slices added
to a thread. Whereas before, the nesting structure of the slices was determined
by the order of B/E calls, it is now determined purely by the timestamps.
This considerably improves modularity. The rule is now: slice y is a subSlice
of x iff y.start >= x.start && y.end <= x.end.

BUG=50

Review URL: https://codereview.appspot.com/6347071