[Telemetry] Don't output site breakdown for reference builds.
Currently we are outputting the exact same lines for both the real build
and the reference build. Since the latter takes precedence, only the reference
build numbers are showing up on the graph.

This patch makes it so we only output the breakdown for the real build, not the
reference build. This seems better than outputting both as that would mean 50
traces for the smoothness_benchmark.

BUG=163520
TEST=None

Review URL: https://codereview.chromium.org/11447026

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@171526 0039d316-1c4b-4281-b951-d872f2087c98