[Telemetry] Fixing credentials login behaviour in case the user is already logged in and using --dont-override-profile.
Currently the page will timeout and fail if it's already logged in, since it doesn't see the login form.
I also added a warning when users specify --dont-override-profile, and a facebook_backend_unittest.py.

Note that this depends on https://codereview.chromium.org/11745002/ to work properly, since currently
the extra quotes around the user-agent string will cause us to be redirected to the wrong version of
accounts.google.com, so we won't correctly detect that we're logged in.

BUG=160268


Review URL: https://chromiumcodereview.appspot.com/11828048

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@176152 0039d316-1c4b-4281-b951-d872f2087c98