Fix a counter drawing bug.
This change fixes a bug with a drawing optimization in CounterTrack.  The
optimization skipped samples when their x-value of the sample would be very
close to that of the previous sample.  Instead the optimization should skip
samples when the x-value is very close to that of the *next* sample.

BUG=136

Review URL: https://codereview.appspot.com/8737043