Allow TimeToObjectInstanceMap to vend different object types
The TimeToObjectMap incorrectly assumed that all the objects vended from it
were of the same type. This patch moves id and parent to the map, but makes
name and category arguments on the idCreated/addSnapshot/idDeleted calls. This
allows, for example, a Frame and Picture with id=0x1000 to be correctly stored
within the map [at different times].