Set a max height on the category filter dialog.
If you have a lot of categories available it's possible for the category
filter dialog to be longer then the available screen height. This makes it
difficult to click the record button. Set a max height of 500 pixels on the
category filter contents (+ 22ish pixels for the button on the bottom) so we
won't overflow the screen on small displays.

BUG=149
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/9067043