Handle objects being deleted properly
We used to ignore 'D' events when they dont have snapshots between them.
However, an object can be deleted and then referenced in an implicit
snapshot. In that case, we get a series of deleted calls, and then
during the implicit id creation, a series of snapshot calls. In that case,
try to extend the lifetime of the deleted object that is most immediately
to the right of the snapshot.