[Telemetry] Fix page_cycler for pages with client side redirects.
Page.disable clears all scripts to evaluate on commit. There is no way
to detect client side redirects because linkedin, for example, performs
its client side redirect after the page has fully loaded.

In order for the page_cyclers to work on pages with client side
redirects, the script to evaluate on commit must execute on the
redirected page.

So to workaround this, we avoid ever calling Page.disable. I don't
expect continuing to listen for notifications here to impact
performance, but if we do see any performance reduction on the bots we
should revert and search for another approach.

BUG=253604
TEST=tools/perf/run_measurement --browser=canary page_cycler tools/perf/page_sets/key_mobile_sites.json --pageset-repeat=1
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/18270002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@209505 0039d316-1c4b-4281-b951-d872f2087c98