[Telemetry] Don't warn if flash is not available when running on non-desktop.
desktop_browser_finder is used even when running cros or android. Since the
test for the ppapi flash binaries was there, it was leading to log spam when
running on those platforms.

This patch moves the check into desktop_browser_backend so that other platforms
will not be bothered.

BUG=260809
TEST=Ran a page cycler on linux
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/19418003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@211861 0039d316-1c4b-4281-b951-d872f2087c98