Move Python setup/tear down logic into Forwarder itself.
Forwarder used to be a pain to setup/tear down across all the various
harnesses.
This CL should hopefully solve these issues by hiding these implementation
details. The host daemon is now killed once the first time that the Forwarder
class is used and the daemon running on the devices is also killed the first
time a port is forwarded for a specific device.

BUG=242846
R=bulach@chromium.org, frankf@chromium.org

Review URL: https://codereview.chromium.org/18086004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@212020 0039d316-1c4b-4281-b951-d872f2087c98