Add an InstantEvent class.
This is the first stage in getting instant event scopes to render. I've created
and instant event class but am still using it as though all of the events are
thread height. The next step will be add the events to trace_model and the
process tracks and have them render at the correct height.

BUG=trace-viewer:156
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/11358044