Extract CounterSeries out of Counter class.
Currently the Counter class does some weird magic to store all of the
counter series information in a set of arrays. This is confusing and hard
to maintain. This CL extracts the series information out to their own
class which is used by the Counters.

BUG=trace-viewer:323
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/11513044