Prevent exception when key is pressed with no selection in list view
The picture ops list view extends list_view's functionality to support
deselecting if a selected element is clicked twice. If a key is pressed
in this case, an exception would be thrown because list_view assumes
there is a selected element.

This patch just checks that this.selectedElement is defined in the key
event listener.

TBR=nduca

Review URL: https://codereview.appspot.com/11643043