Convert flow events to use the interval tree.
This switches the flow event processing to use the interval tree instead
of the sorted array to determine the slices which are onscreen.

I've also fixed the drawArrows code to not take the absolute value of the
slice start distances. This has the side effect that the flow events which would
previously disappear off the screen don't get drawn as they all have a slice
start time for the end event that is before the start time of the beginning
event.

BUG=
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/12797043