Add alpha parameter to FastRectRenderer and pass async property
This change adds a parameter for alpha to the FastRectRenderer and
prioritizes colors with higher alpha for drawing merged rectangles.

The async property of slice tracks is passed to the draw helper methods
that make use of the alpha parameter.

This change also draws labels of async slices in italic font for easier
distinction.

BUG=74
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/13235048