Reland: [Telemetry] Properly handle retries after browser crash.
Previously it was easy for the results to get into a bad state and assert in a
variety of ways.

This ensures we ignore the partial results in the case of a crash.

This reland is tested with GPU functional tests. The main fix was to make all
result types deep copyable instead of just CSV.

BUG=
TBR=kbr@chromium.org

Review URL: https://codereview.chromium.org/138143021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@245788 0039d316-1c4b-4281-b951-d872f2087c98