Clean up style sheet and url handling in tvcm
URLs in style sheets were handled with a lot of ad-hoc code, namely:
- The generator for css had code to inline images to b64 data uris
- check_grit existed to ensure that uris pointed at things that existed
- all this code assumed that images were in src/images

When any of these assumptions were changed, things broke.

This changes the loader system to parse all style sheets encountered. A
style sheet object is introduced that reads the style sheet and resolves any
urls to Image objects, throwing errors if errors are found.

This eliminates the need for check_grit, since the correctnes of the style
sheets is now verified with every test run.