Allow following flow event arcs with shift-arrow keys.
Currently the arrow keys will move to the next slice in a given track. If we're
on a flow event arc we usually want to move to the next item in the flow. This
CL changes the behaviour of the shift-arrow keys to move through flows if you
have a flow event as your starting point.

BUG=trace-viewer:536, trace-viewer:246
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/77780043