Fixed lthi_cats_perf_test
The bug is caused by a wrongly specified file url to import.

Refactored the code a bit to not have url strings written twice in the code.
The problem was that we are looking for a lthi_cats.json file, while the
repository has a lthi_cats.json.gz file. Also in order to parse the gzip file,
we need to get both the gzip importer as well as the trace_event_importer.

BUG=571
R=skyostil@chromium.org

Review URL: https://codereview.appspot.com/111130044