Clean up project path handling
Projects used to get their paths via constructor.

This made subclassing projects really crazy, because the subclassee
had to handle both adding its own paths plus maybe optional additional
paths passed in on the constructor. This was fragile in practice.

This just makes the paths mutable, and adds the right notification handling
so that if you do mutate the paths, the base project finds out.