Remove tvcm/polymer.html: content should include this manually
In preparation for splitting tvcm into pieces, this patch makes
the inclusion of tvcm/polymer.html something up to the toplevel
file building up the final reuslt. This will let the baking parts
of tvcm become py-vulcanize, leaving behind helper functions and a
unit test harness.

This doesn't quite go all the way: the current handling of <script src=> causes
scripts to get processed out of order relative to imports. This makes it
impossible for the generators to put the polymer & platform files before the
imports. The short term fix for this is to make the actual js generator force
include the two raw files in question. A followup patch can clean this up.