Implemented a toggleable container.
This container will help in improving the analysis results
statistics display. Right now, the analysis results uses a tooltip
that is very difficult to read. Also, the fact that it's hidden does not
help with discovering the feature in the first place.

The plan is to add a small button (perhaps a +/- sign) that expands
a row to show the statistics.

R=nduca@chromium.org

Review URL: https://codereview.appspot.com/124460046