Add profiling_controller that manages profilers. Remove StartProfiling & StopProfiling methods from browser.py.
profiling_controller is owned by platform, so this will allow decoupling lifetime of profilers from browser.
This cannot be achieved yet in this patch since we have some
profilers whose lifetime dependent on the browser, e.g: v8, tracing... To do
this, profilers' APIs should be modified for case browser_backend is optional if
they are not dependent on the browsers (e.g: monsoon, tcp_dump...).
For now, StartProfiling will explode if the number of running browsers is not 1.

BUG=

Review URL: https://codereview.chromium.org/511653002

Cr-Commit-Position: refs/heads/master@{#293618}