Add why_imported script
./vulcanize_trace_viewer --report-sizes let us understand where bytes go in the final trace_viewr.html,
but --report-deps gives you such a monstrous graph that its pretty hard to understand causality.

So, instead, once you identify a big module you wish not to be included, use why_imported. Given no arguments,
it'll tell you about all the things in the world that could cause your module to get used. Or, passed a config,
e.g. --config=systrace, you can see what in that config is pulling in the module of concern.

Pass the resulting file to dot:
  why_imported --config=systrace base.rect | dot -Tsvg > /tmp/why.svg