Actually fix scrolling: the interactive_test_runner selector was broken
The reason for https://github.com/google/trace-viewer/commit/9d2402 was that
test scrolling was broken. This was reverted because the fix changed the actual
body to scroll.

The real bug was that the interactive-test-runner had lost its styling due
to a typo during the tvcm move. This caused it to not get sized,
which prevented it from overflowing, which prevented scrolling.

This fixes the styling, and everything works again.