Implement asynchronous filtering
This patch changes the find control to use asynchronous tasks instead
of synchronous blocking to find the matching trace events. This lets the
rest of the user interface to update during an on-going search.

Note that this change by itself does not make the UI completely
responsive during searching; constructing the table of results is
another big blocking task that needs to be chopped up.