Add location class
A location is an object that represents a spatial point on the timeline
stored privately as the xWorld for the x direction, and percentage offsets
of all bounding tracks for the y direction to handle for collapsing tracks.

Will be used later to implement annotations.
Next step will be to add unit tests.