telemetry: Add a method to access tabs by key in TabList.
Previously, the only public method to access tabs from TabList was by index
into an array. The problem with this method is that the array is dynamic, so
there's always the possibility of a race condition causing the index
dereference to explode.

The plan is to migrate the interface entirely over to accessing tabs by key.
For now, add a public method that allows access to tabs by key. Also, make tabs
return the same key as their id property.

BUG=398467, 442546

Review URL: https://codereview.chromium.org/915173012

Cr-Commit-Position: refs/heads/master@{#316530}