Remove one use of PageTest.CanRunForPage.
After this only the startup benchmarks will have a dependency. The goal is to remove CanRunForPage entirely.

session_restore has used CanRunForPage to select one page from the typical_25. The URL is not used because the navigation action is skipped. The only reason it is there is because there is a test when setting up Web Page Replay that the archive has the URL.

BUG=440101

Review URL: https://codereview.chromium.org/808893002

Cr-Commit-Position: refs/heads/master@{#317734}