Fix zoom animation's affectsPanY()
affectsPanY() would always return true because this.goalPanY is never
assigned a value. I believe this.goalFocalPointY was the intended
variable to be used.

This is not a visible issue since currently affectsPanY() is only
used in the viewport to stop active animations if an onScroll event is
invoked. If the zoom animation truly did affect panY the animation would
be cancelled there anyway. Nevertheless I believe it's still worth fixing.

I only found this issue when trying to animate panY for my
NavUsingFindController feature.

Transform animation unit tests' coverage has been improved to test
for this as well.