Fix display of arguments-that-have-objects in single_slice_sub_view
The conversion of sssv to table_buidler lost proper display of complex objects.
This was because the one test for this capability,
instantiate_withSingleSliceContainingIDRef, did not actually validate that the
actual object link was created.

The test was fixed, and another test that uses a simple object in its argument
was also added.

Fixes #812