change annotation getOrCreateView
Previously when constructing an annotation a viewport needed to be
provided. This is a problem because during the process of importing
events a viewport is not available.

This is now changed so that annotations can be created and added to the
trace model without a viewport. A viewport is only necessary when
attempting to retrieve an annotation view (through getOrCreateView(vp)).

R=dsinclair@chromium.org

Review URL: https://codereview.appspot.com/219230043