profiling view saving a file even after cancelling save dialog
When in profiling view, clicking save creates a dialog prompt asking
for a custom filename. When clicking cancel in that dialog, a file
named "false" is still downloaded. This was because the false constant
when added to the document element.download was interpreted as a string.

Also fixed a presubmit formatting error in same file.

BUG=http://crbug.com/470042
R=dsinclair@chromium.org

Review URL: https://codereview.appspot.com/218160043