change location obj viewport to be optional during construction
We need to construct location objects (for RectAnnotation) during
the importing phase, and a viewport is not available during that.

This change allows a location to be constructed without a valid viewport
(using the constructor, not the static functions). In that case a viewport
must be given during toViewCoordinates() instead.

R=dsinclair@chromium.org

Review URL: https://codereview.appspot.com/216280043