[telemetry] Sort imports in Telemetry and its dependents.
As part of the Telemetry public API refactor, I wrote a tool that sorts imports.
The tool will be checked in separately, but here's all the import changes in Telemetry and its dependencies.
It'll be easier to test the refactor tool if we do the import sorting in advance, so it doesn't clutter the diffs for the module moves.

Style guide rule: "Within each grouping, imports should be sorted lexicographically, ignoring case, according to each module's full package path."

BUG=449308
TEST=trybots (tools/telemetry/run_tests; tools/perf/run_tests)

Review URL: https://codereview.chromium.org/1029263003

Cr-Commit-Position: refs/heads/master@{#322211}