Fix GoV to preserve its style and update SingleSliceSubView accordingly
SingleSliceSubView had a forced monospace on it. But, this was mostly to ensure
that args were formatted sanely. This was in turn required because GoV was
clobbering its style element by assigning this.shadowRoot.textContent = ''.
This makes GoV behave sanely wrt Polymer, and then removes the cruft in SSSV.

Review URL: https://codereview.appspot.com/216660043