accept any stableId in nav find string
Previously we would only accept string in the form of
(timestamp)@(pid).(tid)x(zoom)
So now instead of pid.tid, we can accept any stableId.
This will be useful when I implement a way to generate the
find string so that it covers as much of the timeline as
possible. ie. allow a nav string to point to a counter track,
kernel track, etc.
(https://github.com/google/trace-viewer/issues/895)

Also fixed the UI State tests.. it turned out it was never testing for
anything.

R=nduca@chromium.org

Review URL: https://codereview.appspot.com/226320043