Improve memory dump process pane by hiding table when no process is selected
Since the columns shown in the table of the memory dump process pane now
depend on the selected process (more precisely on the attributes of the
memory allocator dumps), only the "Allocator" column is shown when no
process is selected. As a result, the table doesn't stretch the whole
width of the window (because the "Allocator" column has fixed width),
which looks bad.

This patch fixes the issue by hiding the table when no process is
selected (showing a "No process selected" info text instead).

BUG=#849
R=primiano@chromium.org

Review URL: https://codereview.appspot.com/234120043