Added GPU Async Slice parsing support as a plugin.
This CL should be landed in conjunction with the chromium side CL:
  https://codereview.chromium.org/1089873004/

Although this depends on the chromium side changes for the GPU Tracer,
it is written so it behaves as it did before if the argument "gl_name"
is not found. If the argument "gl_name" is found, it utilizes the
gl_name as the trace title, and modifies the SubGroupTitle so that
it differentiates Device and Service categories.

In order to better correlate service and device traces for each
channel, the colors have also been modified so they match.

R=fmeawad@chromium.org, nduca@chromium.org
BUG=https://code.google.com/p/chromium/issues/detail?id=478923

Review URL: https://codereview.appspot.com/233800043