Add memory maps pane to single memory dump analysis sub-views
This patch modifies the single process and global memory dump analysis
sub-views to show a "Memory maps" details pane when a cell in the
"Proportional used memory (mmaps)" column on the "Overview" pane is
selected.

A screenshot of the new pane was added to the tracking bug
(https://github.com/google/trace-viewer/issues/832).

Note that this patch also (1) modifies the memory dump analysis panes to
prevent flexbox shrinking and (2) turns on selection support for all
memory dump analysis tables to make it possible to highlight table rows.

BUG=#849
R=nduca@chromium.org, primiano@chromium.org

Review URL: https://codereview.appspot.com/233520043