Overhaul drawing container perf tests
In order to address #957, we will need to include a perf test that is
async-slice heavy. However, the drawing container perf tests are very
densely coded right now in a way that makes it hard to slot in a different
model. Moreover, the test names are set such that when you click the _10,
you end up also running the _100 test. Thats frownyface.

This switches the drawing container perf test to subclass the PerfTestCase,
which will let us add more subclasses in a followup patch.

Review URL: https://codereview.appspot.com/237430043