Speed up async slice rendering by 4444x
This makes the async slice heavy perf test run in 11.83ms/cycle, down from
48.83ms/cycle, fixing #957.

The first core issue is that the top bar drawing code was flushing the fast
rect renderer every single time we transitioned between a toplevel slice and a
non toplevel slice. So tracks that had a mix of toplevels and non-toplevels
would basically lose all their rendering speed.

The second core issue is that the rects were done as lines interleaved inside
the fastRect renderer code. This meant that we'd be issugin
fillRect/line/fillRect/line over and over again, the exact pattern that makes
canvas unhappy.

The third core issue is that the top bars weren't being fast-rect rendered. The
FRR combines many small rectangles into a single larger one when they are
subpixel in size. The topbar code was not doing this, so when you were zoomed
far out, we drew lines for every single topbar.

A side issue was that the topbar picked up the global alpha of whatever was
drawn previously into the FRR. Once we draw an async slice, alpha would be 0.7,
but before that, it would be 1.0. This meant that if the first slice drawn had
a topbar, that slice's topbar would be 1.0 alpha. But any subsequent topbars
would be drawn with a 0.7 alpha. This fixes #971.

This reorders things quite a bit for better performance:
* The topbars are done as a second pass.
* The topbars go into the fast rect renderer
* The FRR doesn't flush when setYAndHeight is called but it doesn't change

TBR=nduca
BUG=#957,#971

Review URL: https://codereview.appspot.com/232650043