[Android] Speculative fix for perf bot woes.
The N4 and N9 perf bots have been flaky since I landed
https://codereview.chromium.org/1132993004. That CL added three uses of
DeviceUtilsRunShellCommand with check_return=true, which throws an
exception on nonzero exit codes. This was a functional change from the
existing code, unlike the rest of the changes. I'm backing them out as
a speculative change with the theory that an exception is being caught
and silently eaten, causing us to skip whatever follows the RunShellCommand
and subsequently fail.

BUG=489317
TBR=nednguyen@chromium.org,sullivan@chromium.org

Review URL: https://codereview.chromium.org/1148293013

Cr-Commit-Position: refs/heads/master@{#333006}