Expand the set of dashboard modules in catapult.
The purpose of this is to test more things, and to verify that the
parts of the dashboard that depend on the now-added third-party modules
still work.

Most of this code has been copied over from the latest version of the
dashboard, but some of it has been modified to remove internal-only names.

Review URL: https://codereview.chromium.org/1149703006