Implement RAILScore class
The previous RAILScore was more of a function that computed a result dictionary.
This changes the interface around so that RAILScore is an object that remembers
how it was derived, which lets us then pass a RAILScore into a side panel object
for visualization.

Credit for this patch goes to @benshayden.

Review URL: https://codereview.appspot.com/247920043

Patch from Ben Hayden <benjhayden@chromium.org>.