Stop printing logs while running dashboard tests in catapult.
Some of the dashboard tests invoke functions which log errors,
but these are not interesting or indicative of actual errors when
running the test.

By setting the logging level in run_tests.py, this makes it so
no log messages are printed when run_tests.py is run, so the
output is a bit cleaner.

Review URL: https://codereview.chromium.org/1180363002