Change 'subsystem_size' to 'size' in memory dumps
This patch changes the code for calculating memory dump size from
adding a new 'subsystem_size' attribute to overriding the existing
'size' attribute.

The reason for this is that we don't want to expose too many size
metrics, as it would be too confusing. This should be the 'size' that
most people expect given that it's a simple aggregation of the existing
size attribute. We will expose differences (e.g. where a parent is
smaller than the direct sum of its children) with icons/tooltips in the
table.

BUG=#849
R=primiano@chromium.org

Review URL: https://codereview.appspot.com/242550043.