Only shift the power timestamps if the power series is new.
This is a fix to a little bug I found in the code where we would shift the
power timestmaps even if the series was not added because the power track is
already there.  It's unlikely we would ever run into this bug though.

BUG=#1012
R=dsinclair@chromium.org, petrcermak@chromium.org

Review URL: https://codereview.appspot.com/243150043