Add <tr-ui-b-view-specific-brushing-state> element
This patch adds a new element for handling brushing state of arbitrary
views (e.g. analysis sub-views). An element can use the new element by
instantiating it and appending it to itself:

  <div id="some-view-with-specific-brushing-state">
    <tr-ui-b-view-specific-brushing-state
        view-id="unique-view-identifier">
    </tr-ui-b-view-specific-brushing-state>
    <!-- Other child elements -->
  </div>

The state can be retrieved from and pushed to the new element as
follows:

  newStateElement.set(state);
  state = newStateElement.get();

Under the hood, the new state element searches the DOM tree for an
ancestor element with a selectionController field to persist the state.

BUG=#706
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/250810043.