Do not create failures when an expected-fail test starts passing
The d8_compatibility unittests are only temporary until run_d8_tests
is working. These tests run very slowly, and when you make a single
change to improve run_d8_tests, twenty to thirty files suddenly
start passing. This then involves a long run of the compat tests,
and then a manual update of 30 or so expectations.

This patch stops raising errors when an expected-to-fail tests
actually passes. This allows work on run_d8_tests to proceed,
but preservs the regression-prevention part of the compatibility tests.

TBR=nednuyen@chromium.org
BUG=#984

Review URL: https://codereview.appspot.com/247550043