Convert xsrf_token to string before validating token in xsrf.py.
Apparently the fact that self.request.get('xsrf_token') was returning
a unicode object rather than str was causing validation to fail.

This affected the tests run on the bot environment, but not locally,
possibly because a new version of App Engine SDK was in use, but I didn't
verify this.

This should fix dashboard tests (ip_whitelist_test 403 errors) on
the catapult waterfall.

BUG=505849

Review URL: https://codereview.chromium.org/1218103006