Changes the device's power series so that it has to be set directly.
Before, we were lazily initializing the power series so that we weren't
taking up resources with it for the vast majority of traces where no
power data was present.

However, the device track will need to check if any power samples are
present in order to know whether to display. Checking this involves
accessing the power series... which would once again mean that we're
initializing it for all traces.

Instead, we're now forcing the caller to initialize the series.

R=dsinclair@chromium.org

Review URL: https://codereview.appspot.com/255140043.