Add mocha and chai to catapult/third_party.
The versions of mocha and chai used here were copied over from trace-viewer.

Also in the first version of this CL: modify the couple of sample dashboard tests to use the local version of chai and mocha.

This is not strictly necessary for dashboard, since it would be possible to use CDN for chai and mocha, but having copies in third_party has the possible advantages of:
 - No network requests required to fetch libraries when running tests locally
 - Not relying on external CDN

BUG=488609

Review URL: https://codereview.chromium.org/1232703004