Modifies PowerSeries and PowerSamples to hold references to parents.
This is necessary for the multi power sample analysis view. When an
analysis view is rendered, the view only has access to the selected
events - in our case, the power samples. However, in order to generate
the 'power consumed over the course of each frame' graph that we want
for the multi power sample analysis view, we need access to the vsync
timestamps (on the device) which specify when one frame ends and the
next begins. In other words, the power samples need to be able to access
the device, which this CL allows us to do.

BUG=#996
R=dsinclair@chromium.org

Review URL: https://codereview.appspot.com/255320043.