Properly detect presence of globals when installing hooks
During the d8 transition, a bunch of global hooks were patched around
because they don't exist in d8. Especially, window.onerror.

However, the way the code checked for the presence of these features
is by checking for it being truthy. That works for things like global.sessionStorage,
which is by its nature, always truthy.

The problem happens when you come to window.onerror: it exists on the global object,
in a hasOwnProperty sort of way, but it is null. So its present, but falsy. The result
is that we've not been catching uncaught errors in tests ... for two weeks, since
23bcbd9ad29d105f44bf039f8c72005027c31e3e.

This uses tr.isExported, which does the right thing.

BUG=#1106

Review URL: https://codereview.appspot.com/256920043