Make profiling_view_test.recording test handle errors correctly
This is yet-more manual promise handling issues: the test uses a
timeout to click on the recording dialog. This assumes that the
recording dialog came up safely. When it doesn't come up, then it
tries to click on a nonexistant button, which then causes us to
eat the real excepiton in the selection dialog starting up in the
first place.

BUG=#1106

Review URL: https://codereview.appspot.com/253340043