RecordAndCaptureController did not handle errors in showTracingDialog
p2 = promise.then(successFunc, failFunc) does not call failFunc to be called
when successFunc throws. It only causes p2 to reject. This means that if
successFunc throws, and you don't do something to connect p2 to your overall
promise chain, you will lose failures in successFunc.

This was causing us to lose exceptions in the showing of the record selection dialog
on Canary.

BUG=#1106

Review URL: https://codereview.appspot.com/253350043