Implement full text filtering
The current filter text box only matches slice categories and titles.
This makes it tricky to find slices with specific arguments.

This patch implements filtering against the full text contents of
slices. This includes, in addition to the category and title, any
arguments of the slice.

R=dsinclair@chromium.org

Review URL: https://codereview.appspot.com/254250043 .