Build a header component to start replacing header track.
This CL adds a header polymer element that we can use in place
of the header track. I've converted rect_track and space_track
over to use this instead of inheritance.

The heading_track was never really a track, it just added
a header onto the start of the track. This makes it clearer
that the header is just a component and has nothing
to do with the track itself.

R=nduca@chromium.org, nduca@google.com, petrcermak@chromium.org

Review URL: https://codereview.appspot.com/256020043 .