Remove FindAllModules code from tvcm/devserver
This removes the FindAllModules function from the core of tvcm, and from
the callsites in some of the helper/lint scripts that we no longer need.

This function was responsible for our crazy non_module_html_files system,
and so with the core FindAllModules departure, we no longer need that either.
So, its gone, too.

TBR=nednguyen@chromium.org

Review URL: https://codereview.appspot.com/253510043.