Revert "RAILIRFinder finds Responses and Animations."
This reverts commit 2fbf6a1380858445395a9ef841d11eca144e45d1, and
followup commit 59047a53a28553f4e9c296e016c92372b636a1d3.

The issue is that the ir_finder tests use the ir verifier. But, the ir verifier
is a polymer component. That doesn't work because the ir finder and its tests need to work
in vinn/node (browserless): in those environments, we don't have Polymer.

We need to rework this patch so that the verifier is pure js, and the ui for debugging
the results is optional.

TBR=benjhayden@chromium.org

Review URL: https://codereview.appspot.com/256260043.