Change root path for the tracing directories.
Currently all of the trace viewer files assume the directory root is the
tracing/tracing directory and all URLs are relative to there. This isn't great
once we try to re-use things with other projects.

This CL updates all of the code to assume URLs are relative to tracing/ and
fixes up the various scripts so those paths will work.

BUG=#1194

Review URL: https://codereview.chromium.org/1277903004