Implement RAILIR.normalizedCpuEfficiency.
It is an affine interpolation (normalize()) of the sum of the cpuSelfTime of associated ThreadSlices.

After computeNormalizedPain is committed, we could switch this to use that
instead if you want. https://codereview.chromium.org/1282343002

Each IR can say how much CPU could be spent on it without impacting the
efficiency score (minCpuFraction), and how much CPU would need to be spent to
completely trash the efficiency score (maxCpuFraction).

minCpuFraction and maxCpuFraction have default values in RAILIR

minCpuFraction and maxCpuFraction are over-ridden by IdleIR.

https://github.com/catapult-project/catapult/issues/1077
BUG=#1077

Review URL: https://codereview.chromium.org/1283883003