Use pyc-agnostic method of locating App Engine modules
Previously, __file__ was being used to attempt to locate App Engine modules
so they could be executed using subprocess, which was buggy because __file__
can point to .pyc files, which cannot be executed.

This CL introduces a small helper called module_finder, which uses the imp
library, which interacts with the import system to provide the correct path.

BUG=#1310
TEST=dashboard/bin/dev_server && dashboard/bin/deploy

Review URL: https://codereview.chromium.org/1303163004