[systrace] Fix path for prefix.html and suffix.html
prefix.html and suffix.html are relative to systrace.py and
they do not change dynamically. When we use a wrapper in
Android repo, the script_dir will be the wrapper dir. We
want to use systrace_trace_viewer.html in the wrapper dir,
but we still want to use prefix.html and suffix.html in
systrace_dir.

BUG=#1339

Review URL: https://codereview.chromium.org/1319083002