Short-circuits stripe generation when no vsync times exist
Functionally, this CL is a no-op. Previously, we were relying on the
fact that findLowIndexInSortedArray returned an index greater than the
array's length if and only if handed an empty array, in which case it
returns an index of 1. However, handling of this case is much clearer
if we just short-circuit and return an empty list from the function.

Review URL: https://codereview.chromium.org/1323083002