Remove the sunburst and use a tree view for sample frames.
This CL updates the code to use a tree view instead of the sunburst in order
to display stack frames. This makes it a lot easier to view the frames and
see how the time has been used.

I've removed the sunburst_chart and the sortable_table code as they are no
longer used now that sampling_summary_panel has been removed.

BUG=#1233

Review URL: https://codereview.chromium.org/1315843003