Android event log importer.
This importer parses the Android event log,
which can be dumped with "adb logcat -b events -d".
The event log contains information about which
activities are created, resumed and paused.

Adds a simple activity class to the model,
which can be used to represent an Android
activity running in the foreground.

BUG=#1127

Review URL: https://codereview.chromium.org/1319013002