Roll typ to v0.9.4.
This picks up several bug fixes in typ, and will most notably add
a new debugging option (-vvv) that will help with debugging some
telemetry issues.

It is also unfortunate that we have two copies of typ checked in
at this point; we should switch to pulling this in via DEPS, but
this is not the time and place for that change.

R=nednguyen@chromium.org
TBR=nednguyen@google.com
BUG=457446

Review URL: https://codereview.chromium.org/1328053002

Cr-Commit-Position: refs/heads/master@{#347539}