[Dashboard] Use 'std' when parsing list_of_scalar_values if it exists.
Telemetry now uploads a 'std' field which indicates a precomputed
standard deviation, which may use a different statistical algorithm than the
Dashboard's default. Thus, if 'std' is present in a chart JSON trace dict, we
should use it instead of computing it.

PERF SHERIFFS: This may change the width of the error bars displayed on charts.

BUG=#1426

Review URL: https://codereview.chromium.org/1335363003