AnimationIR.normalizedJankPain uses Statistics.timestampsDiscrepancy() on the frame timestamps.
Discrepancy quantifies the uniform-ness of a set of numbers.

If every frame of an animation is the same length, then the timestamps are very uniform and discrepancy is very low, less than 0.05.

If one or more frames are very long, then the timestamps are not very uniform, so
the discrepancy will be higher, maybe up to 0.3.

JankPain contributes to AnimationInteractionRecord.normalizedUserPain.

BUG=#1245

Review URL: https://codereview.chromium.org/1339743002