Add GetURLForAbsFilename to app
This is needed so that we can figure out what the serving location of a file inside the
trace hierarchy is, in order to enable clicking on traces that have been processed by
perf insights.

Unfortunately, doing this requires adding methods on the app object. Previously, the
app object was not something we had subclassed. So first this patch changes us to subclassing
WSGIApplication. Then it adds the path conversion methods.

Yet more unfortunately, we have a bit of a layering puzzle to wrestle with: the core of this patch is just walking through source paths until you find a match to the filename. But, source paths is only known by the app. Okay, so we ask the app for the url. Except, the app doesn't know its URL prefix, e.g. http://blah/, just its routes. Darn. So, then we need to teach the app about its server. Except, the app is created before the server. Bleh.

The final flow is:
- SourcePathHandler knows how to go to url and from url, given
  a source paths list
- App is created, figures out its serving paths, remembers them. Also
  has uninitialized info about its server. At this point, the app cannot respond to path->url requests (or stop server requests).
- Server is created, and tells the app that it now exists. Now, app can respond to path->url requests and stop server requests.

BUG=#1378
R=sullivan

Review URL: https://codereview.chromium.org/1365323002