Fixes a bug where Scheduler::BeginFrame vsyncs were parsed incorrectly
This is a modified reland of 5290486, which incorrectly assumed that all
Scheduler::BeginFrame events had an associated argument object with a
frame time. It turns out that this frame time was added a few months
ago, so any traces older than that with Scheduler::BeginFrame events
were crashing on import due to slice.args.args being undefined.

BUG=#1581

Review URL: https://codereview.chromium.org/1387923003