Style change: Remove unnecessary `os.path.abspath()`.
This is a small, relatively unimportant CL.

In tracing_project, perf_insights_project and dashboard_project we are setting a bunch of path variables which are all relative to the catapult repo root.

As long as the catapult repo path is absolute, then all of the other paths (which are based on it and don't include "..") will also be absolute. Removing the call to os.path.abspath makes the lines a bit shorter.

Review URL: https://codereview.chromium.org/1391543005