Make hovering of analysis links cause their events to brighten
Previously, we had only three color variations to choose from: regular,
brightened, and dimmed. Thus when something was both selected and also
highlighted we couldn't distinguish the two visually. This meant that
hovering analysis links did not cause the events under that link to
brighten.

This adds another color variation: super bright. The HIGHLIGHTED state is
then mapped to this variation.

As part of this, the color palette was moved to use colors instead of
color-like dicts.

R=beaudoin
BUG=#1117

Review URL: https://codereview.chromium.org/1417553005