Add mechanism to display importWarnings that are of interest to user
importWarnings track a new field called showToUser. When true,
the timelineView will show this error to the user.

The RAILIRFinder is then updated s.t. any exceptions that happen inside it are
caught and shown to the user. This ensures that we notice RAILIRFinder
exceptions and try to fix them.

This is a continuation of https://codereview.chromium.org/1418013002/

BUG=#1243

Review URL: https://codereview.chromium.org/1420183002