Actually fix analysis link hovering
This introduces multiple brightening and dimming levels to the color palette.
The presence of an event in any of the brushing state then increments or
decrements an events brightness level. Roughly, the more brushed you are, the
greater your brushing level. When there is a restriction, e.g. due to related
events or find matches, then we start the world at a dim state, whereas without
a restriction, then we start the world at the default state.

This allows an event to be selected, and hovered, at the same time, and get the
right brightening applied. Or, for it to be a non-selected find match, and get
brightened.

Note: The previous commit fa87f4592fc89633e66f5e8d899a1411de2ac16b
was incorrectly titled as fixing analysis link hovering. That patch
was just cleanup in preparation for this patch, but the CL description
was stale at committing time. Whoops.

TBR=beaudoin
BUG=#1117

Review URL: https://codereview.chromium.org/1429863004