Don't discount tracing from missing VM region byte stats
This patch updates the logic for discounting tracing overhead from
memory dumps so that the overhead wouldn't be subtracted from the
relevant byte stats (private dirty and PSS) which are not provided.

For example, given a heavy process memory dump (with VM regions), if
*none* of the associated VM regions has a defined PSS, the tracing
overhead will *not* be subtracted from the total PSS of the VM regions.
On the other hand, if at least one of the regions has a defined private
dirty size, then the tracing overhead *will* be subtracted from the
total private dirty size (even if it wasn't subtracted from the total
PSS).

Rationale: This is to avoid displaying negative values for unsupported
byte stats (e.g. PSS on Mac).

BUG=#1667

Review URL: https://codereview.chromium.org/1430843002