Make UserFriendlyCategory a passive part of the model
PerfInsights introduced the concept of a user friendly category. This is really
useful for summarizing the model in user friendly terms, but as implemented, is
completely external to the model. This means that things like the single event
sub view cannot easily access this informatio to improve information display.

One way of fixing this would be to have chrome auditors iterate every event in
the model and add a userFriendlyCategory property to every single event. The
downside of this is that ufc's are currently dynamically calculated and thus
an approach of string ufc's on events would increase memory consumption.

The alternative that this patch uses is to allow the UFC to be obtained from an
event dynamically. The model tracks a list of UserFriendlyCategory drivers.
Asking for an event's UFC iterates that list until one of the drivers gives a
result. Auditors install UFC drivers. This allows android to have a UFC driver
at the same time as Chrome, for instance.

One awkard part of this patch is that getUserFriendlyCategoryForEvent is on
model, not on event. This is not ideal. Unfortunately, events do not always
have a pointer to their parent. Thus, event cannot get to the model to get the
drivers. This patch tries to split the difference, since fixing the parenting
problem is a pretty long term issue.

BUG=catapult:#1511

Review URL: https://codereview.chromium.org/1412003009