Fix broken rules in html_test_results.html
The html_test_results tagname was corrupted during one of the
big namespace normalization pushes. This caused its rule for display-as-flex
to not be found, which caused overflow to not be triggered on the results
view. This meant that when the results view needed to shrink to acommodate
other parts of the interactive_test_runner element, it wouldn't, and we
would get visual artefacts like in #1731

BUG=catapult:#1731
TBR=dsinclair

Review URL: https://codereview.chromium.org/1421893007