drag_handle_test should clean up its changes even when tests fail
Drag handle test adds things to document.body as part of its testing.
But its cleanup code does the removal only if the tests pass. So if the
tests fail, then there are drag handle tests sitting in the body.

This puts the cleanup in a finally block.

TBR=dsinclair
BUG=catapult:#1731

Review URL: https://codereview.chromium.org/1422053007