Actually fix dragHandle tests in canary
Yet more flexbox changes broke our tests. The draghandle
test rig was expecting a 1 1 auto flexed item to shrink, but it had
no minHeight applied to it. So it woudln't actualy shrink enough.

This adds minHeight, which causes the shrinking to behave as expected.

TBR=beaudoin
BUG=catapult:#1731

Review URL: https://codereview.chromium.org/1409063005