Swap the order of 'effective_size' and 'size' column in the memory UI
This patch changes <tr-ui-a-memory-dump-allocator-details-pane> to show
the effective size column before the size column.

Rationale: Effective size is displayed everywhere in the memory UI (e.g.
in memory dump chart tracks and the memory dump overview pane), so it
should also be the first column on the allocator pane.

BUG=catapult:#1461

Review URL: https://codereview.chromium.org/1407883009