Recompute RAILScore off rangeOfInterest
Currently, selecting a range of the timeline view does not cause the
rail-score-span in the top right corner nor the rail-score-side-panel
to update to reflect the selected range.

Both the rail-score-span and the rail-score-side-panel always reflect the
entire timeline.

But sometimes we just want to see the RAIL score for the IRs in a selected
time range, or narrow the rail-score-side-panel to a selected time range.

This CL makes the TimelineView update the rail-score-span when the
rangeOfInterest changes.

This CL makes the rail-score-side-panel update its rail-score-span and
its table when the rangeOfInterest changes.

In order to facilitate these changes, the RAILScore now takes an optional
Range, and only considers IRs that intersect that range,
if it is not empty.

https://github.com/catapult-project/catapult/issues/1746
BUG=catapult:#1746

Review URL: https://codereview.chromium.org/1419573014