Use the LoadFinished event to find MainFrame LoadIRs
Currently, LoadIR has a parentRoutingId that identifies its parent RenderFrame.

RAIL IR Finder finds the LoadIR's parentRoutingId by finding the
RenderFrameImpl::createChildFrame event that created the loading frame.
RAIL IR Finder drops sub-frame LoadIRs by only producing a LoadIR if its
parentRoutingId is undefined.

Currently, RAIL IR Finder uses the parentRoutingId to determine whether
a Load IR is a MainFrame.

The problem is that createChildFrame events are unreliable.
They might have happened before tracing started.

A better way to find whether a LoadIR is a MainFrame involves using the
WebCore LoadFinished instant event. This event is traced during
RenderFrameImpl::didFinishLoad only if the RenderFrame is a MainFrame.

If the end of a MainFrame Load is captured in the trace, then the LoadFinished
event will always be present.

If the end of a MainFrame Load is not captured in a trace, then we wouldn't
have been able to compute any meaningful metrics for it anyway.

This CL changes RAIL IR Finder to use the WebCore LoadFinished event
as a clearer and more reliable proxy for whether a Load is a MainFrame.

Along the way to facilitate this change, this CL adds explicit access
to the Load IR's renderer process object and renderer main thread object.

Eventually, tracing will represent FrameTrees as first-class objects,
and match Load IRs up with first-class WebFrame objects directly.
But the LoadFinished event is the best we can do for now.

https://github.com/catapult-project/catapult/issues/1298
BUG=catapult:#1298

Review URL: https://codereview.chromium.org/1420273012