Make GuessMetric always return a two-part metric (incl if there is a TIR label)
https://code.google.com/p/chromium/issues/detail?id=556753

This CL changes GuessMetric to *always* return a string with exactly
two slash-separated parts (chart/trace), where if there is an interaction
record label, then it's included as part of the "chart" part.

I expect that this should make the metric strings returned by GuessMetric
work as-is with legacy bisect without having to be rewritten, and they should
also work with recipe bisect IF the revision is after the change from
using "@@" as a separator to using "-" as a separator.

Review URL: https://codereview.chromium.org/1472973004