Add support for customizing memory dump analysis view table columns
This patch finally removes the hacks for coloring used memory and
tracing columns in memory dump analysis sub-views. Instead of modifying
columns after they were created, it is now possible to specify the
MemoryColumn class to be used based on the name of the column.

BUG=catapult:#1461

Review URL: https://codereview.chromium.org/1476923007