Update closure-linter to v2.3.19
This patch updates the Closure Linter library, which is used for
checking JS code style, to the newest version.

Rationale: Several files in the Catapult repository now use ES6 features
which are not supported by the old version of the library (namely line
breaks within for...of statement headers).

BUG=catapult:#1499

Review URL: https://codereview.chromium.org/1556363002