Speed up browser-side HTML test running by 274%
This CL includes a variety of optimizations to the test runner
to improve its speed:

1. All test results widgets are created up front, instead of prepended as we
go. This I think fixes the flake issues that caused us to originally want that
approach, but doesn't come with the usability mess it causes. And, it makes the
layout less destructive each time: prepending is a known slowpath in the layout
engine.

2. I removed the addHTMLOutput validation of every new element having size.
This forced layout at the tune of 30ms per call on a macbook pro. At 1500
tests, this is a hefty bill.

3. I batch N (16 currently) tests into one requestIdleCallback call. rIC is
way too conservative to actually pile things reasonably for our purposes.

Before, the run time was 30.7seconds. Now it is 11.1sec.

TBR=nduca
CC=beaudoin

Review URL: https://codereview.chromium.org/1659793003