Move findInputIRs() to its own file.
Currently, RAILIRFinder contains all of the logic to find RAIL IRs in chrome
traces... in a single file!

This huge file is difficult to deal with, and contains multiple separate
functional units.

Also, framing the (too-large) unit as a class incorrectly encourages attaching
difficult-to-find pieces of state to the ephemeral RAILIRFinder object.

This CL splits out the findInputIRs() function into a separate file.
The function is kept as a function in order to discourage accumulating
difficult-to-find pieces of state.

BUG=catapult:#1940

Review URL: https://codereview.chromium.org/1665443002