Remove checking for Bug entity when posting bisect results.
We have TryJobs that don't have correlated Bug entity.  This is the case when someone sends a try job and manually entering a bug id. On the other hand, we automatically create a Bug entity when alerts are triaged.  Bug entity are used for managing alerts.

Change here is to go ahead and post bisect results if TryJob has a valid bug_id.

BUG=catapult:#2072

Review URL: https://codereview.chromium.org/1743823002