clock sync: Create the ClockSyncManager to connect clock sync markers
This CL is the result of recent discussions around clock sync as well
as http://bit.ly/1TlD0t5.

One good idea that was proposed was that we should structure the
ClockSyncManager in a way so that, if we ever decide to merge the clock
sync pass with the 'shiftTimestampsToZero' pass, we can do so easily. I
think that this can be easily done with the current interface done by
renaming 'ClockSyncManager' to 'ClockManager' (or something like that)
and adding a new method:

  addTimestampToClockDomain(clockDomainId, timestamp)

All other changes required would be internal.

BUG=catapult:#1606

Review URL: https://codereview.chromium.org/1753963002