In start_try_job.GuessBisectBot, lower-case platform string.
In the case of Nexus 5X, it was checking whether the platform string "nexus5X" is in the lower-cased bot name "android-nexus5x". I believe the intent of lower-casing the bot name was to make the check case-insensitive; in order to do this, the platform string must also be lower-cased.

https://bugs.chromium.org/p/chromium/issues/detail?id=595740

Review URL: https://codereview.chromium.org/1806333007