tr-ui-b-table.defaultExpansionStateCallback
Currently, it is ergonomically difficult for UI components to define
when a table row should be expanded. Table requires isExpanded to be
set in the tableRows when the tableRows are set on the table.

This CL adds defaultExpansionStateCallback to allow embedders to dynamically
determine whether a row should be expanded. This allows expansion to be
expressed as an encapsulated heuristic, separate from tableRows construction.

BUG=catapult:#1765

Review URL: https://codereview.chromium.org/1778743002