Modifies model.html to use arrow functions for anonymous functions
This is in an effort to determine whether projects that depend on
trace viewer can handle ES6 features. Our approach is to start using
an ES6 feature in a small but core part of the codebase and see whether
anything downstream breaks. If it does, we easily roll back the change.
Otherwise, we can safely assume that we can use other stable and
supported ES6 features in our code.

BUG=catapult:#1499

Review URL: https://codereview.chromium.org/1358233002