Split update_test_suites to internal and external cron jobs.
This will be a quick fix to give more time for update_test_suites to run.  This does make update job for internal and external run in parallel, which should be ok.

To reduce the bottleneck even further, we can turn this into two taskqueue tasks which can run serially, or use pipeline (but I think we need a generic runner to make it reusable).

BUG=catapult:#2122

Review URL: https://codereview.chromium.org/1820123002