Remove old rail ir verifier view
This view was originally used to aid in debugging the tests for what is now the
UserModelBuilder.

Currently, it's unused.
We had some ideas for a way to use it that wouldn't violate any layers,
but the UserModelBuilder isn't changing in ways that would require
this kind of debugging aid, so it's easier to just remove it.

This is the last "rail ir"-branded thing.

BUG=catapult:#1492

Review URL: https://codereview.chromium.org/1823893002