Move importation of custom-tooltip to static templates.
This is a follow-up to https://codereview.chromium.org/1828733002.

It works without moving the HTML imports to static/report.html and
static/group_report.html, but it is better if in general custom elements
are always imported in the same file where they're used.

BUG=catapult:#2166

Review URL: https://codereview.chromium.org/1824293003