Make the trace button disabling behavior the same as the bisect button.
This is a follow-up to https://codereview.chromium.org/1812073003;
in that CL I changed the bisect button to start as disabled and become
enabled only if /can_bisect confirms that bisect is supported, or if
there's an error. The "trace" button should have the same behavior.

BUG=catapult:#2151

Review URL: https://codereview.chromium.org/1832633002