Visualize LayoutTrees
This CL adds LayoutTreeSnapshot to represent LayoutTrees as currently snapshot
in blink when the disabled-by-default blink.debug.layout category is enabled.

LayoutTreeSnapshot converts the layout tree json into LayoutObjects,
which is already defined.

This CL also adds layout-tree-sub-view, which is just a table to display
some of the properties of LayoutObjects.

Review URL: https://codereview.chromium.org/1759403002