Drop CSS Animations that do not contain any frames.
Currently, the UserModelBuilder is happy to produce AnimationExpectations
for any and all CSS Animations, whether or not the screen actually updated while
the animation was "running".

This breaks the system_health animation metrics, which require animations to
have frame events.

This CL modifies handleCSSAnimations in the user model builder to ignore CSS
Animations that do not have frame events.

This CL does not fix other types of Animations such as Scrolls and Flings.
Those animations are not currently represented in the top25 and key_mobile_sites
page sets that are planned for the system_health benchmark, but those animations
will need to be fixed at some point.

BUG=catapult:#1159

Review URL: https://codereview.chromium.org/1844703002