Memoize frame events while finding CSS Animations
Currently, CSS Animations find frames inefficiently.
It is not uncommon for there to be a dozen or more CSS Animations in a trace, so
it's important for each one to be handled in a reasonably efficient manner.

This CL memoizes the frame events for each renderer process that has a CSS
Animation, to help subsequent CSS Animations find the frame events faster.

This CL does not address other types of AnimationExpectations.

This CL does not alter the UserExpectations in any way, so no test changes.

BUG=catapult:#1159

Review URL: https://codereview.chromium.org/1842963002