Add function GetBattorPathFromPhoneSerial() to find_usb_devices.
Adds a function GetBattorPathFromPhoneSerial() which takes a serial number of a phone and returns that TTY path (e.g. /dev/ttyUSB0) that is used to communicate with the BattOr corresponding to that port. This is necessary for tracing agents to get battery traces from the BattOr. See, e.g. go/battor-systrace, go/clock-sync-design.

Review URL: https://codereview.chromium.org/1823193002