Allow metrics to re-use groupingKeys
Currently, Value retains a reference to the groupingKeys dictionary that is
passed to it, and returns the same reference from asDict().
This can lead to unintended consequences if callers re-use dictionaries.

This CL adds tr.b.cloneDictionary() and applies it to Value's handling of
groupingKeys in order to buffer the impedance mismatch of this pass-by-reference.

This will simplify metrics by allowing them to re-use groupingKeys dictionaries
safely.

Review URL: https://codereview.chromium.org/1856193002