GUID.allocateUUID4() generates version 4 UUIDs.
Currently, GUID generates consecutive integer ids for speed.

TBMv2 Values will be gaining ids soon so that they can reference each other.
However, ValueSets may be serialized and merged, so Value's id scheme needs to
prevent collision between runs.

allocateUUID4() generates a random string UUID for use by Value and ValueSet,
so that Values and their ids can be serialized and deserialized, and will still
be unique even if ValueSets are combined across runs.

GUID.allocate() is renamed to allocateSimple() for clarity. This required updating
a few files, and then the include-order-checker and linter required a few knock-on
changes.

BUG=catapult:#2314

Review-Url: https://codereview.chromium.org/1954713002