[tracing] - Fix subtle bug gathering failures in run_metric.
The nested comprehension grabbing the failures is broken, and just happens to work right now because 'trace' is defined already. To make it clear, make this change:

failures = [f for f in results[t].get('failures', []) for t in traces]

Additionally, the order is switched up. By forcing traces to contain failures, ie.

traces = ['a', 'b', 'c']
for t in traces:
  results['trace'] = {'failures': [t]}

You'll get 3 failures 'c', instead of 'a', 'b', 'c'.

This can all be fixed by changing the nested comprehension to:

failures = [f for t in traces for f in results[t].get('failures', [])]

But overall I found this super hard to read and understand, so patch just uses for-loop and extend.

Review-Url: https://codereview.chromium.org/1987563003