Define DiagnosticMap and a basic Diagnostic hierarchy.
Currently, Value.diagnostics is a free-form dictionary.
Metrics can add anything they want to it!
The problem is that metrics don't know what to add to it.

This CL defines a simple hierarchy of strongly-typed Diagnostics
(Generic, RelatedValueSet, Composition),
with nice APIs like DiagnosticMap and RelatedValueSet.

Each subtype of Diagnostic is defined and registered in its own file,
so anybody can define and register a new kind of Diagnostic.

A followup CL will define a diagnostic-map-view and a registry of diagnostic views
for the different types of Diagnostics.
https://docs.google.com/document/d/1F7A0WRVC-aYHSz_rDnEBzDpjEDJl0_1Kjb9IKrpAhok/edit

In order to support the new RelatedValueSet and Composition diagnostics,
placeholder ValueRef objects are resolved into Values using ValueSet
after all Values are imported from dicts,
so that individual Values can be imported without their entire ValueSet.

The canonical_url diagnostic is unused, so this CL removes it.

Grouping keys are removed from Values as a first step towards refactoring them
into diagnostics, since this CL already significantly changes the Value API.

A followup CL will refactor Value's opt_options and remove 'important'.

The only metric that made extensive use of diagnostics was hazardMetric, which is
being refactored to not use diagnostics anyway:
https://codereview.chromium.org/1992303003

BUG=catapult:#2180

Review-Url: https://codereview.chromium.org/2000063006