[devil] Implement StatDirectory and ListDirectory
Implement StatDirectory and ListDirectory methods in DeviceUtils to
eventually replace the current Stat and Ls (not useful because they
can't take the as_root option).

These new methods are implemented by parsing the output of 'ls -a -l'
on the device. Whether possible and convenient, we attempt to make
parsing strict and return values mirroring those of the standard |os|
and |stat| Python modules.

BUG=chromium:552376

Review-Url: https://codereview.chromium.org/2026723002