Temporary fix for battor_agent failing to run within telemetry.
There is a difficult bug that needs to be fixed in the battor_agent
where input commands can cause a segfault if they are sent within the
first few seconds of executon of the battor_agent.

This patch adds a sleep before commands are send to the battor_agent to
avoid this bug. We need this tempporary patch until the battor_agent issue https://crbug.com/620010 is resolved.

BUG=catapult:#2381

Review-Url: https://codereview.chromium.org/2063913004