Use DIP for scrolling tests.
DIP stands for Device-Independent Pixel. Chrome on CrOS is configured
with IsUseZoomForDSFEnabled. On high-end devices with high DPI display
the return values from __GestureCommon_GetBoundingVisibleRect will be
the number of real pixels, which causes troubles as the scrolling
coordiates are expecting DIP.

The fix here is to convert real pixel values back to DIP by dividing it
with window.devicePixelRatio.

BUG=chromium:599656
TEST=smoothness.top_25_smooth now passes on Pixel

Review-Url: https://codereview.chromium.org/2066573004