Fix incorrect method call in gc_metric.html.
The function findTopmostSlicesRelativeToThisSlice is being called here,
but that function takes three arguments (eventPredicate, callback, opt_this)
Thus, what is happening here is that findTopmostSlicesRelativeToThisSlice
is being called with the function defined (if tr.metrics ... pauses.push ...)
as the eventPredicate, and the callback is being left undefined. Thus, when
the method is called, the eventPredicate is called and has the intended side effect,
and then since there's no return statement in the eventPredicate, it then iterates
through into the subslices, effectively traversing the entire subslice tree and
calling the predicate on it.

Since this file has extensive tests which all pass, I am presuming that this is the
intended behavior. Thus, this CL changes this call to iterateAllDescendents,
the function that is intended to be used for this behavior.

(ulan, benjhayden: am I correct that the behavior is as intended despite the
incorrect method call, or is there actually a functional bug in this code?)

Review-Url: https://codereview.chromium.org/2062273004