Plumb IterationInfo from telemetry to a diagnostic
For now, IterationInfo contains only those fields that are absolutely necessary
for results2.html: storyName, label, timestamp, storySetRepeatCounter,
storyRepeatCounter.

Values can become separated from their ValueSets, and ValueSets can contain
Values from different iterations, and the current IterationInfo is only a
hundred bytes or so, so for now, this diagnostic will be duplicated on every
Value.

This diagnostic will eventually gain more fields and also be displayed on the
dashboard and unified with the existing value metadata system.
For now, only the storyName is replicated between IterationInfo and the existing
value metadata.

If and when this diagnostic grows too large to be replicated on every Value, then
we can choose how to combat the bloat using strategies that are being worked out
in the design doc:

https://docs.google.com/document/d/12KvpnDUFUuCEgId7H-QL5PoRcsp99CXdqbZc-EOrpH8/edit#

BUG=catapult:#2405

Review-Url: https://codereview.chromium.org/2055043002