Plumb IterationInfo through TracingControllerBackend instead of RunMetric
Currently, IterationInfo is plumbed through RunMetric. However, this means that
when metrics are re-computed by passing the trace files through run_metric, then
IterationInfo is not available.

This CL passes IterationInfo through TracingControllerBackend so that it is baked
into the trace so that it is available even when using run_metric.

BUG=catapult:#2180

Review-Url: https://codereview.chromium.org/2090183002