Implement BarChart.isStacked
Currently, our d3-based BarChart is similar to gcharts' "column chart",
d3-based with brushing but without stacking.

Eventually, we would like to have a chart library in tracing/ui/charts/
that is as featureful and expressive as we need it to be,
but have an API based on gcharts
both to help users transition from gcharts-based demos to tracing-charts-based
UIs, as well as to help us minimize the amount of design work that we need to do.
The primary differences between tracing/ui/charts/ and gcharts will be that
tracing-charts is vulcanizable, brushable, and integrates with the rest of
tracing/ui/ and tracing/value/ui/.

For now, we just need a stacked bar chart to facilitate specific
custom value-set-view customers and their Composition diagnostics,
with an API that is a step towards a gcharts-alike tracing/ui/charts/ library.

BUG=catapult:#2180

Review-Url: https://codereview.chromium.org/2099533002