Fix diagnostic analysis-links
Currently, clicking analysis-links in diagnostic-span just throws an obscure error.
This CL interposes a Value-aware BrushingStateController in between the
analysis-links and the Value-unaware BSC, and makes value-set-table handle the
requestSelectionChange event by selecting the selected Value.

This CL makes analysis-link.onClicked_ stop the click event from propagating,
so that clicking an analysis-link in value-set-table will not select the table cell,
since value-set-table is just going to select another cell right away in response
to the requestSelectionChange event.

Review-Url: https://codereview.chromium.org/2143293002