This CL is the foundation of the NetLog viewer web app including base functionality to import a NetLog file. With this CL the web app has seven fully functioning tabs: Import, Events, Proxy, Timeline, DNS, Sockets, and Cache.
The main change in this CL is the use of html imports in index.html to glue all of the elements from the imported documents onto the main document. Other changes include deprecating the import view's use of jstemplate and removing features that have no place in this new web app (capture and export).

Please look over the README.md file for instructions on how to run the web app.

Modifications to come in later CLs:
1) Deprecating communication with the browser in browser_bridge.js
2) Updating issues with the currently functional tabs. (Sockets tab remove "Close" and "Flush" buttons, Cache tab fix "Explore cache entries" link, and Proxy tab remove "re-apply settings button")
3) Test infrastructure and test cases (this approach has been deemed reasonable because of how this project is mostly copied-over code from Chromium and because the current plan is to integrate the tests while receiving feedback).
4) Enabling sdch, spdy, Alt-Svc, and Quic tabs by removing their dependencies on the jstemplate.

BUG=chromium:472699
NOPRESUBMIT=true

Review-Url: https://codereview.chromium.org/2090743002