Add mock XHR to instantiation test for test-picker.
This reproduces the problem where the subtest menu does not show up,
but doesn't fix it yet. Awkwardly, the tearDown needs to be commented
out in order for this to work, since the user interaction that loads
the menu happens after the test finishes. It should be possible to
automate this, but I'm not sure how, and this quick change gives us
what we need to debug.

BUG=catapult:#2285
NOTRY=true
NOPRESUBMIT=true

Review-Url: https://codereview.chromium.org/2151403002