Copy a couple NetInternals test files from Chromium's source code:
https://chromium.googlesource.com/chromium/src/+/1be518f2fe0c5e815de8ecd64496088d7b13471e/chrome/test/data/webui/net_internals/

These are just a few test files that will be modified in subsequent
CLs in order to test using Catapult's test infrastructure. The goal
is to have one basic test working to ensure that it is possible
to use Catapult's infrastructure as a way of adding tests for this
directory.

BUG=chromium:472699
NOPRESUBMIT=true

Review-Url: https://codereview.chromium.org/2178043004