systrace: Force systrace to use adb from the devil
run_systrace.py is slow if there's an adb in $PATH that's newer than the one
in devil.

First of all, run_systrace.py needs an adb in $PATH to begin with otherwise
it complaints. So I've had the adb from my android_sdk/platform-tools/ which
I use with my android devies in $PATH.

The issue happens when run_systrace.py initially starts to use the adb from
my $PATH but then code execution goes to devil.android device_utils, it tries
to use the adb from devil, this results in adb server being killed and restart.
Causing a slow down of around 10 seconds in every run.

To work around this, we have run_systrace.py start to use the adb from devil
so that the entire run uses just one version of adb (the one in devil).

Signed-off-by: Joel Fernandes <agnel.joel@gmail.com>