Fix testPersistentShell.
Sometimes adb returns a blank line between two separate calls to
pshell.RunCommand(); sometimes it doesn't. Why it does this is unknown;
however, I did check and there is no place currently using the persistent
shell where this would make a difference. Modified the test to pass
whether or not there is that extra blank line.

BUG=catapult:#2574

Review-Url: https://codereview.chromium.org/2191283003