[Values] Change the diagnostics type of each Numeric's sample to DiagnosticMaps instead of Diagnostics
In LAWM (https://github.com/catapult-project/catapult/issues/2588),
we find a case where we would like to attach both a generic
diagnostic (timestamp information) and a composition diagnostic
(breakdown) to values produced by loadingMetric. This makes that
possible by mapping samples to DiagnosticMaps, instead of single
Diagnostics as before.

This is necessary to land the initial Lawm report as-is
(https://codereview.chromium.org/2197773003).

BUG=catapult:#2588

Review-Url: https://codereview.chromium.org/2219733003