Make MultiSampleSubView not use selection.getEventsOrganizedByBaseType
The MultiSampleSubView uses getEventsOrganizedByBaseType to subset the
selection to find the samples. But, this can be implemented by a filter
step, too.

This patch moves to the filter approach in order to make it be that only
analysis_view is using getEventsOrganizedByBaseType. This will allow a followup
patch to remove that function from the EventSet and further customize it for
the analysis view only.

R=lpy@chromium.org
BUG=catapult:#2672
NO_DEPENDENCY_CHECKS=true

Review-Url: https://codereview.chromium.org/2240313003