[Telemetry] Also merge across tir_labels for TBM
This CL updates the summarization behavior of chart JSON to summarize twice:
once by name and tir_label (which is the current behavior), and once by name
only (so that merging across tir_labels will occur). The reason we want to do
this is to fill in the summary values for story sets with story keys, which we
abuse the tir_label field to store.

It also updates the merging code for SummarizableValues, which previously
assumed that merging would only occur for lists of values which all had the
same tir_label.

Currently, in the presence of IR labels, we only output these summaries:
* Per story/tir_label
* Per tir_label

With this change, we will additionally output, for TBMv2 benchmarks:
* Per story, across all tir_labels
* Per value, across all stories and tir_labels

BUG=chromium:624522

Review-Url: https://codereview.chromium.org/2142293002