Import polymer in a single place and remove d3 hacks
All key bootstraps just import tracing/ui/base/polymer.html which then initializes Polymer to shadow mode before importing. As part of this, I removed our dependence on webcomponents (sometimes lite, sometimes not!) because we're Chrome only.

But, you will notice these ugly polymer preload and postload scripts. This is a temporary workaround for #2698, in which PyVulcanize causes us to reorder scripts relative to imports.

Another thing "fixed" in this CL is the early- and incorrect import of D3 added during the Polymer conversion. D3 didn't work because its "generic" module loader that is in d3-min.js gets confused by being inlined (I'm a little confused why). The "better hack" I replaced the early loading with is to trick the d3 module loader into thinking that there's an AMD module loader present, via a preload and postload script. In a world where #2698 was fixed, we could have these be inline in the d3.html bootstrap.

BUG=catapult:#2580

Review-Url: https://codereview.chromium.org/2267503002