Pass in custom options to Systrace agents
This CL refactors the Systace agent API to take in a custom config
object tailored to the object. In doing so, this CL:

- Updates Systrace's StartAgentTracing() to take in custom options
- Updates profile_chrome's StartAgentTracing() as well
- Updates the Telemetry call that calls Systrace
- Updates the unit tests accordingly

Note that profile_chrome still needs to be followed to follow Systrace's
controller API. This will be in a separate follow-up CL.

Because so much code is moved between files, this CL looks
artificially large.

BUG=catapult:#2690

Review-Url: https://codereview.chromium.org/2276263003