[Clock sync] Use clock sync manager to convert trace event timestamps
This is usually a no-op because we always try to use the Chrome clock
domain as the model clock domain. However, without actually getting
a model time transformer for the trace event importer, we never verify
that all of the clock domains are connected.

To make telemetry's integration test of tracing on reference Chrome
binary passing, this CL also fixes how trace importers handle clock
sync events based on telemetry's workaround of using console.time
& console.timeEnd for record clock sync on old chrome binary (see
https://github.com/catapult-project/catapult/blob/cb0d9d7c3b28514bb03d9cc1f1e7bc5cd441bd56/telemetry/telemetry/internal/platform/tracing_agent/chrome_tracing_agent.py#L153)

This is based on charliea@'s original work in https://codereview.chromium.org/2203033002/

BUG=catapult:#2547

Review-Url: https://codereview.chromium.org/2334753002