[value-set-table] Move Show all to the right.
Currently, the Show all checkbox is on the left, so it is the first thing that users
see. However, it is currently always disabled because no metrics generate
RelatedValueMap, RelatedValueSet, or Composition diagnostics.

Even when metrics do annotate value relationships with diagnostics, confronting
users with the decision of whether to Show all should probably not be higher
priority than selecting a reference column or searching for a value name.
Starting the controls row with a search box guides users towards the Name
column of the table proper, and is a more familiar control type that signals that
users can safely ignore the controls row until they need it.

Review-Url: https://codereview.chromium.org/2344463002