Merge NumericValue into Histogram.
Pour one out for Value, y'all.
The Value system was a really neat architecture. It made good sense.
But when the TBM2 rubber hit the road, it turned out that the only Value
subclass that was needed was NumericValue.
And in order for merging NumericValues to be predictable, they need to specify
reasonable ranges, and, well, be Histograms.
The TBM2 UIs should handle the scalar statistics.
If all metrics always produce NumericValues that contain Histograms, it's just
more ergonomic if those are the same object.
This aligns with UMA histograms, which have names and bin configurations and
descriptions.

I considered splitting Histogram into a base/ Histogram without name and
diagnostics, and a MetricHistogram with them, but the only non-metric client of
Histogram (multi_sample_sub_view) needs diagnostics (though not name).
So let's keep a monolithic Histogram for now. We can split it if and when a
fault-line emerges in the use-cases. This means that tracing/ui will depend on
value/ (when lpy lands it), but that's already been the case since ScalarNumeric
was moved to value/.

This will invalidate existing results2.html again.

A follow-up patch will move ScalarNumeric to base/ with the Unit system, and
delete NumericBase.
Another follow-up will rename ValueSet, value-set-table, sourceValues, etc.

BUG=catapult:#2685

Review-Url: https://codereview.chromium.org/2334233003