Move scalar-span#significance emoji after #content.
value-set-table right-aligns its cells, which contain scalar-spans.

Make it easy to vertically scan significance emoji by putting
the constant-width emoji to the right of the variable-width content.

If there are any use-cases that want the emoji to the left of the content,
then maybe scalar-span can put #significance after #content when the scalar-span
right-aligned, but before #content otherwise?

Review-Url: https://codereview.chromium.org/2341763003