Use histogram for multi event details view.
Currently we are using multi event details table when we select multi events
that have the same type, but it will freeze the UI when the number is too many
and we try to scroll, besides the table will make us hard scroll to the runtime
call stats table. This patch replace the table with histogram.

The main benefit is it won't freeze the UI while we can still have details about
the events we select.

BUG=catapult:#2800

Review-Url: https://codereview.chromium.org/2346363002