[tracing] Change breakdown view dimension upon pressing left/right arrows
This patch improves the ease-of-use of the new Heap Details UI by
allowing the user to change the dimension of the breakdown view (right
sub-pane). As a result, once the breakdown view is focused, all
navigation can be performed using the keyboard very easily (with no need
to use the Tab key):

  Arrow up/down:    select child node
  Arrow left/right: change breakdown dimension ("stack frame" or "object
                    type")
  Enter:            enter child node
  Backspace:        go back to parent node

As part of this patch, <tr-ui-b-tab-view> is modified to properly update
the UI when a sub-view is selected programatically.

BUG=catapult:#2750

Review-Url: https://codereview.chromium.org/2387393002