Revert "Enforce One True Brace Style in eslint."
This reverts commit 5884cc83313c68df197ac32cd2486eee0723f966.

Reverting this after discussions between eakuefner@ and me re: the idea
that there are probably too many existing violations of this style to
require developers to do the migration for us by enforcement of a linter
error.

As an example, https://codereview.chromium.org/2409683002/ generated
about 250 violations of these rules.

I think we should revert this while we try to see how close a few well-
crafted regexps can get us to having this be the standard in our
codebase, at which point we can turn this on.

BUG=catapult:#2819

Review-Url: https://codereview.chromium.org/2410543003